actually generate a C struct for the dispatch
This adds a new loader gl_struct.py, since the loader part has to be
modified, too. Copying that file seems not like the best approach, but
glad's current design is not ideal for what we are trying to do.

This is work-in-progress, the GL version and extension aviability is still
stored in global variables. Also, the loader internally uses some global
variables, so it is not yet threadsafe.